+++ src/main/kotlin/org/wisp/gatesawakened/activeGateIntel/ActiveGateIntel.kt
@@ -1,14 +1,13 @@
 package org.wisp.gatesawakened.activeGateIntel
 
 import com.fs.starfarer.api.campaign.comm.IntelInfoPlugin
+import com.fs.starfarer.api.impl.campaign.intel.BaseIntelPlugin
 import com.fs.starfarer.api.impl.campaign.intel.misc.BreadcrumbIntel
 import com.fs.starfarer.api.ui.SectorMapAPI
 import com.fs.starfarer.api.ui.TooltipMakerAPI
-import com.fs.starfarer.api.util.Misc
-import org.wisp.gatesawakened.Gate
+import org.wisp.gatesawakened.*
 import org.wisp.gatesawakened.constants.Tags
-import org.wisp.gatesawakened.di
-import org.wisp.gatesawakened.isActive
+import org.wisp.gatesawakened.midgame.Midgame
 
 
 class ActiveGateIntel(val activeGate: Gate) : BreadcrumbIntel(null, activeGate) {
@@ -32,14 +31,17 @@ class ActiveGateIntel(val activeGate: Gate) : BreadcrumbIntel(null, activeGate) 
     override fun createSmallDescription(info: TooltipMakerAPI, width: Float, height: Float) {
         info.addImage(di.settings.getSpriteName("illustrations", "dead_gate"), width, 10f)
         info.addPara(
-            "There is an active Gate here.",
+            "There is an active Gate in this system.",
             10f
         )
+        if (Midgame.wasQuestCompleted) {
+            info.appendPara("${BaseIntelPlugin.BULLET}You have %s activation codes left.", 0f, Common.remainingActivationCodes.toString())
+        }
     }
 
     override fun createIntelInfo(info: TooltipMakerAPI, mode: IntelInfoPlugin.ListInfoMode?) {
         super.createIntelInfo(info, mode)
-        info.addPara("There is an active Gate in %s", 0f, Misc.getHighlightColor(), activeGate.starSystem.baseName)
+        info.appendPara("${BaseIntelPlugin.BULLET}In %s", 0f, activeGate.starSystem.baseName)
     }
 
     override fun getSortTier(): IntelInfoPlugin.IntelSortTier = IntelInfoPlugin.IntelSortTier.TIER_1